+++ controller/src/main/java/com/linbit/linstor/dbcp/migration/k8s/crd/K8sCrdMigration.java
@@ -1,10 +1,10 @@
 package com.linbit.linstor.dbcp.migration.k8s.crd;
 
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
 import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
 
 @Retention(RUNTIME)
 @Target(TYPE)
@@ -13,4 +13,6 @@ public @interface K8sCrdMigration
     String description();
 
     int version();
+
+    boolean isInitial() default false;
 }